- added types_windows.go file which includes the exports for windows functions since these functions are global per package - added windows version of the diskspace module - cleaned up some of the sys call functions
